WEATHERDATA-3 - Add localization Added new en-US and hu-HU resources, also kept the invariant one Localization is realized via the WPF Localize Extension Added language selector which changes the current culture Connected resource strings to the UI Rearranged the JSONMOdels all into separate classes as some of them grew
