+++ DarkSkyWPF/DarkSkyWPF/Properties/Resources.en-US.resx
