+++ DarkSkyWPF/DarkSkyWPF/packages.config
@@ -7,4 +7,6 @@
   <package id="System.Reflection.Emit" version="4.3.0" targetFramework="net461" />
   <package id="System.Windows.Interactivity.WPF" version="2.0.20525" targetFramework="net461" />
   <package id="Unity" version="5.8.13" targetFramework="net461" />
+  <package id="WPFLocalizeExtension" version="3.3.1" targetFramework="net461" />
+  <package id="XAMLMarkupExtensions" version="1.5.1" targetFramework="net461" />
 </packages>
